+++ frontend-prototype/src/App.tsx
@@ -1,0 +1,7 @@
+import StudioApp from "./studio/StudioApp";
+
+function App() {
+  return <StudioApp />;
+}
+
+export default App;